Removing unused sign and renaming module
