+++ src/protocols/two_party_ecdsa/lindell_2017_keygen/party_one.rs
@@ -15,6 +15,7 @@
 */
 
 use ::BigInteger as BigInt;
+
 use ::Point;
 use ::EC;
 use ::PK;
@@ -33,6 +34,8 @@ use cryptographic_primitives::commitments::traits::Commitment;
 use cryptographic_primitives::hashing::hash_sha256::HSha256;
 use cryptographic_primitives::hashing::traits::Hash;
 
+// TODO: remove the next line when unit test will be done
+#[allow(dead_code)]
 pub struct FirstMsgCommitments {
     pub pk_commitment: BigInt,
     pk_commitment_blind_factor: BigInt,
