+++ src/protocols/two_party_ecdsa/lindell_2017_keygen/party_two.rs
