+++ src/protocols/two_party_ecdsa/mod.rs
@@ -14,5 +14,5 @@
     @license GPL-3.0+ <https://github.com/KZen-networks/multi-party-ecdsa/blob/master/LICENSE>
 */
 
-pub mod keygen_lindell17;
-pub mod sign_lindell17;
+// Fast Secure Two-Party ECDSA Signing by Yehuda Lindell (https://eprint.iacr.org/2017/552.pdf).
+pub mod lindell_2017_keygen;
